pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.64.0",
    "asyncio-throttle>=1.0.0",
    "black>=23.0.0",
    "griffe>=1.12.0",
    "openai>=1.0.0",
    "pgvector>=0.2.0",
    "pydantic>=2.0.0",
    "pydantic-ai==0.0.14",
    "pydantic-settings>=2.10.1",
    "pypdf2>=3.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "python-magic>=0.4.27",
    "ruff>=0.1.0",
    "streamlit>=1.28.0",
    "supabase>=2.0.0",
    "tenacity>=8.2.0",
    "tiktoken>=0.5.0",
]
